removed unnecessaray entry from ready files list
The addition of another entry of the form "ready_{gen}_{gen}" is redundant and leads to errors as the file is never specifically created. It has been working most of the time because it is the same name as the "ready_{gen}_{ind}" file. However, it breaks when there are more generations than individuals per generation.